Rename package form subcomponents to sections.
