+++ plugin/src/main/java/net/dzikoysk/funnyguilds/config/sections/LocationConfiguration.java
@@ -1,4 +1,4 @@
-package net.dzikoysk.funnyguilds.config.subcomponents;
+package net.dzikoysk.funnyguilds.config.sections;
 
 import eu.okaeri.configs.OkaeriConfig;
 import org.bukkit.Location;
